feat: add validator
This change adds code to define a validator. This validator can be
used within an adapter to invoke a set of ValidatiorFunctions, which
are functions dedicated to validate the adapter resource.

By calling `Validate`, all the functions passed as an argument will
get evaluated and a final ValidationResult will be returned, which
can be used to condition the operator behaviour.

Signed-off-by: David Moreno García <damoreno@redhat.com>